+++ es7s_commons/gradient.py
@@ -3,7 +3,6 @@
 #  (c) 2023 A. Shavykin <0.delameter@gmail.com>
 # ------------------------------------------------------------------------------
 
-from logging import getLogger
 import re
 import typing
 from abc import ABCMeta, abstractmethod
@@ -13,7 +12,7 @@ from dataclasses import dataclass
 from typing import TypeVar
 
 from pytermor import RGB
-
+from .common import logger
 
 _T = TypeVar("_T")
 
@@ -24,7 +23,7 @@ class GradientPoint:
     col: RGB
 
     def __post_init__(self):
-        getLogger(__package__).debug(f"Created {self!r}")
+        logger.debug(f"Created {self!r}")
 
     def __repr__(self):
         return f"{self.__class__.__name__}[pos={100*self.pos:8.4f}%, {self.col}]"
@@ -38,7 +37,7 @@ class GradientSegment:
 
         col_mid = self._interpolate_2p(self.p_left, self.p_right, pos_mid)
         self.p_mid: GradientPoint = GradientPoint(pos_mid, col_mid)
-        getLogger(__package__).debug(f"Created {self!r}")
+        logger.debug(f"Created {self!r}")
 
     def interpolate(self, pos: float) -> RGB:
         if pos <= self.p_mid.pos: