+++ CHANGES.rst
@@ -12,6 +12,16 @@ Version list
 ---------
 - ...
 
+1.4.0 (Dec 23)
+----------------
+- 🌱 NEW: `termstate.terminal_state` context manager
+- 💎 REFACTOR: `logger` instantiating
+- 🐞 FIX: dependencies
+- 🐞 FIX: `progressbar` component
+- 🐞 FIX: removed debug instruction
+- 🐞 FIX: `RenderableColor` signature
+- ❌ REMOVAL: `joincoal` `isempty` `filtere` `filterev` -> to pytermor
+
 1.3.0 (Nov 23)
 ----------------
 - 🌱 NEW: `common.bcs` `common.lcm`